create delelet trip

+++ app/api/trip/[id_user]/[id_trip]/route.js
@@ -3,6 +3,8 @@
 import Trip from '@/models/Trip';
 import { connectDB } from '@/lib/db';
 import { NextResponse } from 'next/server';
+import User from "@/models/User";
+import bcrypt from "bcryptjs";
 
 export async function GET(req, { params }) {
   const { id_user, id_trip } = await params;
@@ -32,8 +34,10 @@ export async function PUT(req, { params }) {
   return NextResponse.json(updatedTrip, { status: 200 });
 }
 
+// # DELETE /trip/:id_user/:id_trip
 export async function DELETE(req, { params }) {
   const { id_user, id_trip } = await params;
+  const password = await req.json();
   await connectDB();
 
   const trip = await Trip.findById(id_trip);
@@ -41,6 +45,14 @@ export async function DELETE(req, { params }) {
 
   const userExists = trip.user.some(u => u.id_user === id_user);
   if (!userExists) return NextResponse.json({ message: 'User not in this trip' }, { status: 403 });
+
+  const userAdminExists = trip.user.some(u => u.type === 'admin');
+  if (!userAdminExists) return NextResponse.json({ message: 'User not Admin in this trip' }, { status: 403 });
+
+  const user = await User.findOne({ _id : id_user });
+  if (!user || !(await bcrypt.compare(password, user.password))) {
+    return NextResponse.json({ message: "รหัสผิด" }, { status: 401 });
+  }
 
   await Trip.findByIdAndDelete(id_trip);
   return NextResponse.json({ message: 'Trip deleted successfully' }, { status: 200 });
